ai_service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
motor
cachetools
textblob